Added the return params into the report-service because all of the filters actually stem from there.
